Rename debug_filter.dart to development_filter.dart
